requirements.txt
pydantic==2.5.3
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
typing-extensions==4.9.0
aiohttp==3.9.1
tqdm==4.66.1